Game is playable, missing a move
Missing the move that the Rook and the King exchange possitions if they are both on their initial positions. Plus, missing menus and interaction with the player, and I could make an AI option letter. Used the Free Low Res Chess asset from the Unity Asset store for the Piece and Chessboard models.